added some tests to pages and contact controllers

+++ app/controllers/pages_controller.rb
@@ -1,5 +1,5 @@
 class PagesController < ApplicationController
-  skip_before_filter :authenticate_user!, :only=>[:home,:show]
+  skip_before_filter :authenticate_user!, :only => [:home, :show]
   skip_before_filter :verify_login
   #before_filter :superusers_only, :except=>[:home,:show]
 
@@ -25,11 +25,10 @@ class PagesController < ApplicationController
         render "pages/partials/construction"
       else
         @page = Page.find_by_name(page_name)
-        @page ||= Page.find(page_name)
+        @page ||= Page.find(page_name) rescue nil
 
         if @page.nil?
-          @content = "<h1>Page Not Found.</h1>"
-          @title = "NOT FOUND"
+          redirect_to "/", alert: "Page Not Found"
         else
           @content = @page.content
           @title = @page.title
@@ -66,43 +65,45 @@ class PagesController < ApplicationController
     if admin_user?
       @page = Page.new(params[:page])
       if @page.save
-        redirect_to page_url(@page.id)
+        redirect_to page_url(@page.id), :notice => "Page created."
       else
         render :new
       end
     else
-      redirect_to root_path, :alert => "You are not authorized to view this page"
+      redirect_to root_path, :alert => "You are not authorized to create this page"
     end
   end
 
   def update
     if admin_user?
-    @page = Page.find(params[:id])
-    @page.update_attributes(params[:page])
+      @page = Page.find(params[:id])
+      @page.update_attributes(params[:page])
 
-    if params[:commit] == "Update & Continue Editing"
-      render params[:raw]=='y' ? :raw_edit : :edit
-    elsif @page.name == "HOME"
-      redirect_to "/"
+      if params[:commit] == "Update & Continue Editing"
+        render params[:raw]=='y' ? :raw_edit : :edit
+      elsif @page.name == "HOME"
+        redirect_to "/"
+      else
+        redirect_to page_url(@page.name.downcase)
+      end
+      flash.notice = "Page Updated."
     else
-      redirect_to page_url(@page.name.downcase)
-    end
-    else
-      redirect_to root_path, :alert => "You are not authorized to view this page"
+      redirect_to root_path, :alert => "You are not authorized to update this page"
     end
   end
 
   def destroy
-    @page = Page.find(params[:id])
-    if @page.destroy
-      redirect_to pages_path, :notice=> "This Page was destroyed."
+    if admin_user?
+      @page = Page.find(params[:id])
+      if @page.destroy
+        redirect_to pages_path, :notice => "This Page was destroyed."
+      else
+        redirect_to pages_path, :alert => "Error deleting page: " + @page.name
+      end
     else
-      redirect_to pages_path, :alert=> "Error deleting page: " + @page.name
+      redirect_to root_path, :alert => "You are not authorized to delete this page"
     end
-
   end
 
-  def five_secrets
 
-  end
 end
